Split functions + tests into separate files
Having a single src file with an associated spec file was becoming
unruly.
