+++ src/index.js
@@ -1,5 +1,5 @@
 // Import all action handlers
-import ielv from './ielv';
+import ielv from './ielv/main';
 
 // Export all action handlers
 export { ielv };
